Optimize runtime sidecar injection logic
